Add files via upload
Textu[r]al meditation
On the poetic level, Laurie Anderson invites her vireader immediately into her “innerst inner” - innermost insides - her flesh and dream: “This is my dream body. The one I used to walk around. In my dreams.” Yet, the animated pencil sketch of hers in the video invites us to imagine it, since we don’t see the body, and rather the close-ups of the drawn face. The close-up of the plain texture of the paper coincides with the last word dreams as the camera fleetly moves through this emptiness/ textu[r]ality of the paper as if diving in. This videoverse is an anacrusis, it comes before titles, before the story begins. The intonation of the reader separates each phrase. Can one write them in the form of a poem? Does the author want us to restore the text in some form on the textu[r]al void? 

This is my dream body
The one I used to walk around
In my dreams.

Does the textu[r]al void in silence become a space from where a subconscious experience transforms into a figure of speech that draws a new, acoustic, image of the interiors that have no more carnal but textu[r]al body? What is this verse without the sketch? What is verse without this rough texture of the paper? It could have been any imaginary body, very present, very anthropomorphic if we were to imagine it out of the verse alone. But the animated sketch on the paper gives us the countour of the body if only to immediately deceive us into the texture where it dissolves. 
And so begins the drawing of an image emerging from the texture rather than from a memory and visions, which become contours of this image emerging in the eye of the vireader. 
The movie is overflowing with blurred and blurring extensions of a still visions: the snowy road with orange blinking lights of the unmoving tow car,  that draws the sight into the plain texture of wet-white-bright-ness; the ice skater from a home video that looks freezing into a still image, as if redirecting the linear movement in any direction the eye can envision into the textu[r]e of some wight-on-white, as the words and vision dissolve in the animated matter of the old recording disintegrating into sheer noise signals. 
The moving distortion of the lady’s sketch in the beginning of the vipoem creates a similar effect of the animated matter which in simultaneity with her words “I love you forever” merge into an inner vibration of that perpetuates (and has always preceded?) the body that experiences its memories as a return to the matter that vibrates within the animated body. Such an effect is achieved through renarrating the memory into images that, by means of video waves made visible in those distorting effects, dissolve into vibrations. The poet seems to inquire the matter about the matter of love and makes it speak to herself through the images of her memories. 
Laurie Anderson literally captures the “gap” “between the moment that is expiring and the one that is arising” as a body on the animated image becomes more and more a shadow before it dissolves into the acoustic image of “nothing”: this image is constituted of a darkness with the shapes of visual forms looming inside it and emerging slowly into a portrait of a face in front of the closed eyes. As the images emerge from the animated vibrations of the matter (that texture of the page?) in the form of still appearances, the poet walks the vision not as much through particular images, but rather through physical experiences of vision. Like in a meditation, she tells the vireader what is going on with their eyes: now they close, and now they open, and as they open something begins. And it always begins after a plunge into nothingness. The view upon the opening of the eyes is the barren trees against yellow sky - the sky in between the light of the day and the light of the night. The beginning starts from an already living yet still nowhere and nothing. This poetic prelude sets the tune of how one should listen to the story about learning to love posthumously, upon how to live inside where there is no body, but the matter, the texture that comes into the body.